pyproject.toml
[project]
name = "hpc-perf-monitor"
version = "0.1.0"
description = "Performance regression detection service for HPC middleware"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "pydantic>=2.5.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "gitpython>=3.1.40",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.0",
    "aiofiles>=23.2.1",
    "asyncssh>=2.14.0",
]

[project.scripts]
hpc-perf-monitor = "hpc_perf_monitor.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto" 